fix(osc): compatible with cloud odc code, fix ScheduleTaskParameter deserialize… #3202
